refactor(app): hoist the contract types so a slice is actually extractable
Structure review, finding 18, plus finding 17's last misfiling.

The documented "085 layering precedent" — the shell computes the view model,
the FEATURE declares its types — is a workaround for the no-`@/app` rule, and
it inverts the dependency it exists to protect. Measured before the change:
`features/simulator/repo-deps.ts` had 8 consumers, 4 of them shell; `form.ts`
23 with 2 shell; `pins.ts` 10 with 2; `features/editor/RepoPicker.tsx` 3 with
2. Deleting `features/simulator/` broke six shell modules — so the stated goal
of the cross-feature ban, that a slice stays extractable, did not hold.

`src/types/` now holds the contracts only — `LoadedRepo`, `RepoDep`,
`RepoDepsView`, `RepoConnectOffer`, `RepoPickerRow`, `RepoPickerView`,
`FormState`, `PinnedTest` — and both sides import DOWNWARD. Everything that
BUILDS or interprets them stayed in the slice: a type is a contract, a function
is behaviour, and only the contract moved. 35 files rewritten; every affected
import was already type-only, so there is no runtime change at all.

`src/types/` needed no lint-config change, exactly as predicted: the shared
override is written as "everything under src/ that is not a feature or the
shell", so a new folder is inside the boundary by construction. The
`oxlint-boundaries` test confirms it — every file in the new folder gets the
shared bank.

Finding 17's third item lands with it: `app/use-repo-deps.ts` ->
`features/simulator/`. Verified it reads no shell state — every import is
shared-layer or the slice's own — and it sat in `app/` purely because App calls
it, which is the colocation weakness the research report names.

What remains of shell -> feature is now composition (the shell mounting feature
panels, which is its job) plus two hooks calling feature FUNCTIONS. The
type-level inversion is gone.

968 app tests, 326 cli tests, lint and typecheck green.
