+++ packages/app/src/app/use-repo-picker.ts
@@ -10,9 +10,9 @@
  * inherited-config probe — stays in `use-repo-load`, untouched.
  */
 import { useEffect, useMemo, useRef, useState } from "react";
-import type { RepoPickerView } from "@/features/editor/RepoPicker";
 import { listUserRepos, probeConfigFile, repoNote, type UserRepo } from "@/platform/github-repos";
 import { useSyncedReset } from "@/hooks/use-synced-reset";
+import type { RepoPickerView } from "@/types/repo";
 
 /** Rows shown (and probed) at once — the picker surfaces recent work, and
  *  every probe behind it is a real API request. Type to narrow. */
