+++ packages/app/src/features/simulator/use-repo-deps.ts
@@ -14,17 +14,12 @@
  */
 import { useCallback, useMemo, useRef, useState } from "react";
 import { TREE_LISTING_PLATFORMS } from "@/data/host-tokens";
-import {
-  EMPTY_REPO_DEPS,
-  type LoadedRepo,
-  type RepoDep,
-  type RepoDepsView,
-  repoDepsOfFile,
-} from "@/features/simulator/repo-deps";
+import { EMPTY_REPO_DEPS, repoDepsOfFile } from "./repo-deps";
 import { useLatestRef } from "@/hooks/use-latest-ref";
 import { loadEngine } from "@/platform/engine-chunk";
 import { loadRepoFile, loadRepoTree } from "@/platform/run";
 import { causedErrorMessage } from "@/lib/errors";
+import type { LoadedRepo, RepoDep, RepoDepsView } from "@/types/repo";
 
 /** Fetch cap: each package file is one request. The commonest repos fit; the
  *  view counts what the cap dropped rather than pretending it covered all. */
